+++ dspace/modules/additions/pom.xml
@@ -341,7 +341,11 @@
          <artifactId>mockito-inline</artifactId>
          <scope>test</scope>
       </dependency>
-
+      <dependency>
+         <groupId>org.springframework</groupId>
+         <artifactId>spring-test</artifactId>
+         <scope>test</scope>
+      </dependency>
    </dependencies>
 
 </project>
